Remove duplication of HTTP.{do, _getHosts} implementations
