+++ src/common/types/http.ts
@@ -1,12 +1,15 @@
+import Defaults from 'common/lib/util/defaults';
 import Platform from 'common/platform';
+import BaseRealtime from 'common/lib/client/baserealtime';
 import HttpMethods from '../constants/HttpMethods';
 import BaseClient from '../lib/client/baseclient';
 import ErrorInfo, { IPartialErrorInfo } from '../lib/types/errorinfo';
 
 export type PathParameter = string | ((host: string) => string);
 export type RequestCallbackHeaders = Partial<Record<string, string | string[]>>;
+export type RequestCallbackError = ErrnoException | IPartialErrorInfo;
 export type RequestCallback = (
-  error?: ErrnoException | IPartialErrorInfo | null,
+  error?: RequestCallbackError | null,
   body?: unknown,
   headers?: RequestCallbackHeaders,
   unpacked?: boolean,
@@ -25,15 +28,6 @@ export interface IPlatformHttp {
   supportsAuthHeaders: boolean;
   supportsLinkHeaders: boolean;
 
-  _getHosts: (client: BaseClient) => string[];
-  do(
-    method: HttpMethods,
-    path: PathParameter,
-    headers: Record<string, string> | null,
-    body: unknown,
-    params: RequestParams,
-    callback?: RequestCallback
-  ): void;
   doUri(
     method: HttpMethods,
     uri: string,
@@ -43,6 +37,11 @@ export interface IPlatformHttp {
     callback?: RequestCallback
   ): void;
   checkConnectivity?: (callback: (err?: ErrorInfo | null, connected?: boolean) => void) => void;
+
+  /**
+   * @param error An error returned by {@link doUri}’s callback.
+   */
+  shouldFallback(error: RequestCallbackError): boolean;
 }
 
 export class Http {
@@ -67,7 +66,17 @@ export class Http {
   }
 
   _getHosts(client: BaseClient) {
-    return this.platformHttp._getHosts(client);
+    /* If we're a connected realtime client, try the endpoint we're connected
+     * to first -- but still have fallbacks, being connected is not an absolute
+     * guarantee that a datacenter has free capacity to service REST requests. */
+    const connection = (client as BaseRealtime).connection,
+      connectionHost = connection && connection.connectionManager.host;
+
+    if (connectionHost) {
+      return [connectionHost].concat(Defaults.getFallbackHosts(client.options));
+    }
+
+    return Defaults.getHosts(client.options);
   }
 
   do(
@@ -78,7 +87,65 @@ export class Http {
     params: RequestParams,
     callback?: RequestCallback | undefined
   ): void {
-    this.platformHttp.do(method, path, headers, body, params, callback);
+    /* Unlike for doUri, the presence of `this.client` here is mandatory, as it's used to generate the hosts */
+    const client = this.client;
+    if (!client) {
+      throw new Error('http.do called without client');
+    }
+
+    const uriFromHost =
+      typeof path === 'function'
+        ? path
+        : function (host: string) {
+            return client.baseUri(host) + path;
+          };
+
+    const currentFallback = client._currentFallback;
+    if (currentFallback) {
+      if (currentFallback.validUntil > Date.now()) {
+        /* Use stored fallback */
+        this.doUri(method, uriFromHost(currentFallback.host), headers, body, params, (err, ...args) => {
+          if (err && this.platformHttp.shouldFallback(err as ErrnoException)) {
+            /* unstore the fallback and start from the top with the default sequence */
+            client._currentFallback = null;
+            this.do(method, path, headers, body, params, callback);
+            return;
+          }
+          callback?.(err, ...args);
+        });
+        return;
+      } else {
+        /* Fallback expired; remove it and fallthrough to normal sequence */
+        client._currentFallback = null;
+      }
+    }
+
+    const hosts = this._getHosts(client);
+
+    /* see if we have one or more than one host */
+    if (hosts.length === 1) {
+      this.doUri(method, uriFromHost(hosts[0]), headers, body, params, callback);
+      return;
+    }
+
+    const tryAHost = (candidateHosts: Array<string>, persistOnSuccess?: boolean) => {
+      const host = candidateHosts.shift();
+      this.doUri(method, uriFromHost(host as string), headers, body, params, (err, ...args) => {
+        if (err && this.platformHttp.shouldFallback(err as ErrnoException) && candidateHosts.length) {
+          tryAHost(candidateHosts, true);
+          return;
+        }
+        if (persistOnSuccess) {
+          /* RSC15f */
+          client._currentFallback = {
+            host: host as string,
+            validUntil: Date.now() + client.options.timeouts.fallbackRetryTimeout,
+          };
+        }
+        callback?.(err, ...args);
+      });
+    };
+    tryAHost(hosts);
   }
 
   doUri(
